ensure the tz info is kept between cron runs
This fixes a bug wherein the tz information was lost in the
NextableCron. Rather than trying to maintain UTC everywhere internally,
we can just use the fact that jodatime DateTime objects do keep TZ info,
and so we don't need to force UTC everywhere (jodatime does the heavy
lifting when comparing datetimes).