default to TASK_RUNNING status for new tasks
By setting this default, we can reconcile if we don't receive any future
TaskStatus updates from Mesos. The consequence of this is that if the
message to launch the task is dropped, we will atleast find out on the
reconcile; right now, we assume that Mesos got the message (and thus,
we'll hear back in the future), which isn't an assumption we can hold.

Also improves the verbosity of logs, to track down the source of some
messages.