when deserializing, 0 is OK for disk and cpus
This was causing the serializer and deserializer tests to be flaky. I
don't understand how they *ever* pass now that I see the problem? When
deserializing, we were using the mesos_task_cpu mesos_task_disk as values
for cpu and disk respectively if either the cpu value was not set, or it
was set to 0. However, the default value for cpus and disk in the constructor
of BaseJob is 0; thus, when a BaseJob with the default constructor value
for cpu and disk was serialized, that value would be changed to 0.1
upon deserialization.

The solution is to respect 0 as an OK value when deserializing.